Revise the README
This version of the readme is much shorter and simpler than my initial idea. The reason is, if someone is really interested in using or running our software, they need to be directed to the wiki anyways. Perhaps a later update could include build and run instructions, but any instructions we put here will need to be tested on machines other than the quad to ensure accuracy.