Release 1.1 (#196)
* Initial draft, with quite a few TODOs remaining

* Revise the README

This version of the readme is much shorter and simpler than my initial idea. The reason is, if someone is really interested in using or running our software, they need to be directed to the wiki anyways. Perhaps a later update could include build and run instructions, but any instructions we put here will need to be tested on machines other than the quad to ensure accuracy.

* Add I2C interface and GPS node

* Feature: basic safety manager (#176)

* Implement hello-world package safety_manager

* Create service interfaces for safety manager

* Add stubs for test client and server

* First draft of TestServer

* First draft of TestClient

* Add simple tests for TestClient and TestServer

* Create basic safety manager with dummy logging

* safety manager now logs some info about events

* simple tests for the safety manager

* Add watchdog schematic (#180)

* Add basic sidecar design notes

* Add schematic

* Automatic testing with GH Actions (#182)

* Add automatic testing workflow (version 1)

* Fix syntax error

* Skip packages and run on pushes

* Rosdep will now ignore errors

* Install colcon-common-extensions and vcstool

* Install colcon mixins, as in ROS dockerfile

* Removed troublesome packages

On doing a docker build, the mpc_controller and
test_trajectory_following packages were causing rosdep to throw
errors. After talking with Will, it seems we don't actually need
these packages, so they have been deleted.

* Use docker for build and test

* Don't run on pushes

* Remove *_nodes packages from testing

* Retest up to 8 times on failure

* Add sleep time to safety manager test

Co-authored-by: Avery Bainbridge <axb200157@utdallas.edu>
Co-authored-by: Will Heitman <Will.Heitman@UTDallas.edu>
Co-authored-by: Avery <11591624+a-bainbridge@users.noreply.github.com>